Rename tir_to_wir -> wir_build, extract codegen/ module
Following rustc's naming convention (rustc_mir_build pattern):
- tir_to_wir.rs + tir_to_wir/ -> wir_build.rs + wir_build/ (TIR -> WIR translation)
- emit.rs + component.rs moved from tir_to_wir/ to new codegen/
- component_gen.rs, wasm_builder.rs, wasm_postprocess.rs moved to codegen/
- lib.rs Phase 11 split into Phase 11 (wir_build) + Phase 12 (codegen)
