+++ wado-compiler/src/codegen/component.rs
@@ -14,5 +14,5 @@ use crate::wir::WirModule;
 
 /// Build a Wasm Component from a core module and project metadata.
 pub fn build_component(project: &Project, core_module: &[u8], _wir: &WirModule) -> Vec<u8> {
-    crate::component_gen::build_component(project, core_module)
+    super::component_gen::build_component(project, core_module)
 }
